#110 [REFACTOR] 라우트 관련 코드를 route.js로 분리
